requirements.txt
# Core dependencies
setuptools>=65.0.0
pymongo>=4.0.0
pyyaml>=6.0.0
pandas>=1.5.0
schedule>=1.1.0
python-dotenv>=0.19.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950